Add election preparation

+++ ElectronicVoting.Common/Domain/BaseEntity.cs
@@ -4,6 +4,6 @@ namespace ElectronicVoting.Common.Domain
 {
     public class BaseEntity
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
     }
 }
