+++ Validator.Domain/Table/ChangeView/PreElectionVote.cs
@@ -7,6 +7,6 @@ public class PreElectionVote :BaseEntity
     public int Round { get; set; }
     public bool Decision { get; set; }
     public string TransactionId { get; set; }
-    public int ApproverId { get; set; }
+    public Int64 ApproverId { get; set; }
     public virtual Approver Approver { get; set; }
 }
